requirements.txt
ansicolors==1.1.8
docker==6.1.3
h5py==3.8.0
matplotlib==3.6.3
numpy==1.24.2
pyyaml==6.0
psutil==5.9.4
scikit-learn==1.2.1
jinja2==3.1.2
pytest==7.2.2
datasets==2.12.0
requests==2.31.0
pymysql2
